requirements-test.txt
shuup
pytest==5.4.1
pytest-cov==2.4.0
pytest-django==3.8.0
flake8>3,<4
isort>4.3,<5

-e .
